Well it works, and I want to fight Mr. Python himself.
Using PyQt and PyQtGraph
https://pypi.org/project/PyQt5/
http://www.pyqtgraph.org/

I managed to get it drawing some can data. It isn't expandable to more bytes. I slacked off, and that's on me. I wanted to have something to show though. This is an ABSOLUTE bodge job that is held together with duct tape and force of will. It isn't fast enough for live data, but it's faster than matplotlib. I think. Maybe. Everything is chaos. I want to burn my code in a fire. Scrolling is awful. I hate multithreading, I hate python, and this is a mess.

BUT IT DOES KINDA GRAPH. AND IT MOVES.